+++ R/class.R
@@ -12,11 +12,13 @@
 #'   * An S3 class wrapped by [new_S3_class()].
 #'   * A base type, like `logical`, `double`, or `character`.
 #'
-#' @param constructor The constructor function. This is optional, unless
-#'   you want to control which properties can be set on constructor.
+#' @param constructor The constructor function. Advanced use only.
 #'
-#'   A custom constructor should always conclude by calling `new_object()`
-#' @param validator A function taking a single argument, the object to validate.
+#'   A custom constructor should call `new_object()` to create the R7 object.
+#'   The first argument, `.data`, should an instance of the parent class. The
+#'   subsequent arguments are used to set the properties.
+#' @param validator A function taking a single argument, `self`, the object
+#'   to validate.
 #'
 #'   The job of a validator is to determine whether the object is valid,
 #'   i.e. if the current property values form an allowed combination. The
@@ -38,7 +40,6 @@
 #'   allowing greater flexibility.
 #' @return A object constructor, a function that can be used to create objects
 #'   of the given class.
-#' @order 1
 #' @export
 #' @examples
 #' # Create an class that represents a range using a numeric start and end
@@ -65,12 +66,12 @@
 #'     start = "numeric",
 #'     end = "numeric"
 #'   ),
-#'   validator = function(x) {
-#'     if (length(x@start) != 1) {
+#'   validator = function(self) {
+#'     if (length(self@start) != 1) {
 #'       "@start must be a single number"
-#'     } else if (length(x@end) != 1) {
+#'     } else if (length(self@end) != 1) {
 #'       "@end must be a single number"
-#'     } else if (x@end < x@start) {
+#'     } else if (self@end < self@start) {
 #'       "@end must be great than or equal to @start"
 #'     }
 #'   }
@@ -85,7 +86,7 @@ new_class <- function(
     parent = R7_object,
     properties = list(),
     constructor = NULL,
-    validator = function(x) NULL) {
+    validator = NULL) {
 
   check_name(name)
 
@@ -96,6 +97,13 @@ new_class <- function(
          "`parent` must be an R7 class, S3 class, or base type, not %s.", class_friendly(parent)),
        call. = FALSE
      )
+  }
+
+  if (!is.null(constructor) && !is.null(parent)) {
+    check_R7_constructor(constructor)
+  }
+  if (!is.null(validator)) {
+    check_function(validator, alist(self = ))
   }
 
   # Combine properties from parent, overriding as needed
@@ -120,9 +128,16 @@ new_class <- function(
   object
 }
 
-can_inherit <- function(x) is_base_class(x) || is_S3_class(x) || is_class(x) || is.null(x)
+check_R7_constructor <- function(constructor) {
+  if (!is.function(constructor)) {
+    stop("`constructor` must be a function", call. = FALSE)
+  }
 
-is_class <- function(x) inherits(x, "R7_class")
+  method_call <- find_call(body(constructor), quote(new_object))
+  if (is.null(method_call)) {
+    stop("`constructor` must contain a call to `new_object()`", call. = FALSE)
+  }
+}
 
 #' @export
 print.R7_class <- function(x, ...) {
@@ -153,6 +168,70 @@ str.R7_class <- function(object, ..., nest.lev = 0) {
   cat("\n")
 
   if (nest.lev == 0) {
-    str_list(props(object), ..., prefix = "@", nest.lev = nest.lev)
+    str_nest(props(object), "@", ..., nest.lev = nest.lev)
   }
+}
+
+can_inherit <- function(x) is_base_class(x) || is_S3_class(x) || is_class(x) || is.null(x)
+
+is_class <- function(x) inherits(x, "R7_class")
+
+# Object ------------------------------------------------------------------
+
+#' @param .parent,... Parent object and named properties used to construct the
+#'   object.
+#' @rdname new_class
+#' @export
+new_object <- function(.parent, ...) {
+  class <- sys.function(-1)
+  if (!inherits(class, "R7_class")) {
+    stop("`new_object()` must be called from within a constructor")
+  }
+
+  args <- list(...)
+  nms <- names(args)
+
+  missing_props <- nms[vlapply(args, is_missing_class)]
+  for(prop in missing_props) {
+    args[[prop]] <- prop_default(class@properties[[prop]])
+  }
+
+  object <- .parent %||% class_construct(class@parent)
+  attr(object, "object_class") <- class
+  class(object) <- setdiff(class_dispatch(class), "ANY")
+
+  for (nme in nms) {
+    prop(object, nme, check = FALSE) <- args[[nme]]
+  }
+  validate(object)
+
+  object
+}
+
+#' @export
+print.R7_object <- function(x, ...) {
+  str.R7_object(x, ...)
+  invisible(x)
+}
+#' @export
+str.R7_object <- function(object, ..., nest.lev = 0) {
+  cat(if (nest.lev > 0) " ")
+  cat(obj_desc(object))
+
+  if (typeof(object) != "S4") {
+    bare <- object
+    attributes(bare) <- NULL
+    str(bare, nest.lev = nest.lev + 1)
+  } else {
+    cat("\n")
+  }
+
+  str_nest(props(object), "@", ..., nest.lev = nest.lev)
+}
+
+#' Retrieve the R7 class of an object
+#' @param object The R7 object
+#' @export
+object_class <- function(object) {
+  attr(object, "object_class", exact = TRUE)
 }