+++ tests/testthat/test-class-spec.R
@@ -17,6 +17,8 @@ test_that("can work with R7 classes", {
 })
 
 test_that("can work with unions", {
+  text <- new_class("text", character)
+  number <- new_class("number", double)
   klass <- new_union(text, number)
   expect_equal(as_class(klass), klass)
 
@@ -59,7 +61,7 @@ test_that("can work with base types", {
 
   klass <- as_class("character")
   expect_equal(class_type(klass), "R7_base")
-  expect_equal(class_dispatch(klass), c("R7_object", "character", "ANY"))
+  expect_equal(class_dispatch(klass), c("character", "R7_object", "ANY"))
   expect_equal(class_register(klass), "character")
   expect_equal(class_desc(klass), "<character>")
   expect_equal(class_construct(klass, "x"), "x")
@@ -102,7 +104,7 @@ test_that("can work with S3 classes", {
   expect_equal(as_class(klass), klass)
 
   expect_equal(class_type(klass), "R7_S3")
-  expect_equal(class_dispatch(klass), c("R7_object", "ordered", "factor", "ANY"))
+  expect_equal(class_dispatch(klass), c("ordered", "factor", "R7_object", "ANY"))
   expect_equal(class_register(klass), "ordered")
   expect_equal(class_desc(klass), "S3<ordered/factor>")
   expect_equal(class_construct(klass), ordered(numeric()))
@@ -121,13 +123,13 @@ test_that("can work with R7 classes that extend S3 classes", {
   Date2 <- new_class("Date2", parent = Date, properties = list(x = "numeric"))
 
   expect_equal(class_type(Date2), "R7")
-  expect_equal(class_dispatch(Date2), c("Date2", "R7_object", "Date", "ANY"))
+  expect_equal(class_dispatch(Date2), c("Date2", "Date", "R7_object", "ANY"))
   expect_equal(class_register(Date2), "Date2")
 
   obj <- Date2(x = 1)
   expect_equal(obj_type(obj), "R7")
   expect_equal(obj_desc(obj), "<Date2>")
-  expect_equal(obj_dispatch(obj), c("Date2", "R7_object", "Date", "ANY"))
+  expect_equal(obj_dispatch(obj), c("Date2", "Date", "R7_object", "ANY"))
   expect_equal(class_inherits(obj, Date2), TRUE)
 })